+++ ClusterFinder/HoughLineFinder.cxx
@@ -96,7 +96,7 @@ void cluster::HoughLineFinder::produce(art::Event& evt)
 
   mf::LogVerbatim("Summary") << std::setfill('-') << std::setw(175) << "-" << std::setfill(' ');
   mf::LogVerbatim("Summary") << "HoughLineFinder Summary:";
-  for(int i = 0; i<ccol->size(); ++i) mf::LogVerbatim("Summary") << ccol->at(i) ;
+  for(size_t i = 0; i<ccol->size(); ++i) mf::LogVerbatim("Summary") << ccol->at(i) ;
 
   evt.put(ccol);
